+++ todos-auth-fluent/templates/error.mustache
@@ -1,0 +1,7 @@
+{{>head}}
+<body>
+    <h1>Error</h1>
+    <h2>{{statusCode}}</h2>
+    <p>{{message}}</p>
+</body>
+</html>
