+++ todos-auth-fluent/templates/login.mustache
@@ -8,7 +8,7 @@
     <label for="password">Password</label><br/>
     <input type="text" id="password" name="password"/><br/>
     <br>
-    <input type="submit" value="Submit" />
+    <input type="submit" value="Login" />
     </form>
     <p>
     <a href="/signup">Sign Up</a>
