+++ todos-auth-fluent/templates/signup.mustache
@@ -10,7 +10,7 @@
     <input type="text" id="email" name="email"/><br/>
     <label for="password">Password</label><br/>
     <input type="text" id="password" name="password"/><br/>
-    <input type="submit" value="Submit" />
+    <input type="submit" value="Sign Up" />
     </form>
     </p>
 </body>
